We have a movie! And the commentary track too!
This isn't quite ready for release yet, but I wanted to get the scripts
into git now just in case something happens to my hard drive.

The commentary track especially needs a lot of work. I need to take a
hatchet to the dialogue before it's even remotely suitable for release.

The main track just needs some tweaks to the song lyrics and another QC
pass or two before I'll be comfortable with the quality.